1.0 Release

+++ src/main/java/net/brifboy/levelup/repo/UserDBInteraction.java
@@ -1,7 +1,6 @@
 package net.brifboy.levelup.repo;
 
 import net.brifboy.levelup.model.User;
-import net.brifboy.levelup.service.UserClaclulations;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,15 +11,21 @@ import java.util.List;
 @Repository
 public class UserDBInteraction {
 
+    private static final Logger logger = LoggerFactory.getLogger(UserDBInteraction.class);
     @Autowired
     UserRepository userRepository;
 
     public void deleteUsers(List<User> users) {
-        userRepository.deleteAll(users);
+        if (!users.isEmpty()) {
+            userRepository.deleteAll(users);
+            logger.info("Deleted users from DB");
+        }
+
     }
 
     public void saveUser(User user) {
         userRepository.save(user);
+        logger.info("Saved or Updated user in DB");
     }
     public User getUserFormUserIdAndGuildId(long userid, long guildid) {
        return userRepository.getUserFromIdAndGuildId(userid, guildid);
